requirements.txt
PySide6
GitPython